+++ CHANGELOG.md
@@ -14,6 +14,10 @@
   Account.`get_message_by_id` may return `None` in this case.
 
 ## API-Changes
+- Remove bitflags from `get_chat_msgs()` interface #4022
+  C interface is not changed.
+  Rust and JSON-RPC API have `flags` integer argument
+  replaced with two boolean flags `info_only` and `add_daymarker`.
 
 
 ## 1.107.1